refactor: Standardize account bottom sheets and relocate files
