+++ presentation/src/main/java/com/london/presentation/feature/account/components/loggedInContent.kt
@@ -1,4 +1,4 @@
-package com.london.presentation.shared.accountComponent
+package com.london.presentation.feature.account.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -13,8 +13,8 @@ import androidx.compose.ui.unit.dp
 import com.london.designsystem.component.Divider
 import com.london.designsystem.theme.NovixTheme
 import com.london.presentation.R
-import com.london.presentation.feature.accountinfo.account.AccountContract
-import com.london.presentation.feature.accountinfo.account.AccountUiState
+import com.london.presentation.feature.account.AccountContract
+import com.london.presentation.feature.account.AccountUiState
 import com.london.designsystem.R as dsR
 
 @Composable
