+++ presentation/src/main/java/com/london/presentation/feature/account/rating/RatingCategory.kt
@@ -1,4 +1,4 @@
-package com.london.presentation.feature.accountinfo.rating
+package com.london.presentation.feature.account.rating
 
 import androidx.annotation.StringRes
 import com.london.presentation.R
