+++ presentation/src/test/java/com/london/presentation/feature/account/account/AccountViewModelTest.kt
@@ -1,4 +1,4 @@
-package com.london.presentation.feature.accountinfo.account
+package com.london.presentation.feature.account.account
 
 import com.google.common.truth.Truth.assertThat
 import com.london.domain.AppPreferencesService
@@ -8,6 +8,8 @@ import com.london.domain.language.AppLanguage
 import com.london.domain.theme.AppTheme
 import com.london.domain.usecase.GetAccountDetails
 import com.london.domain.usecase.authentication.AuthenticationUseCase
+import com.london.presentation.feature.account.AccountViewModel
+import com.london.presentation.feature.account.ActiveBottomSheet
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
@@ -59,7 +61,8 @@ class   AccountViewModelTest {
         coEvery { accountDetailsUseCase.invoke() } returns AccountInfo(1, "", "")
 
         // When
-        viewModel = AccountViewModel(appPreferencesService, authenticationUseCase, accountDetailsUseCase)
+        viewModel =
+            AccountViewModel(appPreferencesService, authenticationUseCase, accountDetailsUseCase)
         testDispatcher.scheduler.advanceUntilIdle()
 
         // Then
@@ -74,7 +77,8 @@ class   AccountViewModelTest {
         coEvery { accountDetailsUseCase.invoke() } returns AccountInfo(1, "", "")
 
         // When
-        viewModel = AccountViewModel(appPreferencesService, authenticationUseCase, accountDetailsUseCase)
+        viewModel =
+            AccountViewModel(appPreferencesService, authenticationUseCase, accountDetailsUseCase)
         testDispatcher.scheduler.advanceUntilIdle()
 
         // Then
@@ -271,7 +275,8 @@ class   AccountViewModelTest {
         coEvery { accountDetailsUseCase.invoke() } returns AccountInfo(1, "", "")
 
         // When
-        viewModel = AccountViewModel(appPreferencesService, authenticationUseCase, accountDetailsUseCase)
+        viewModel =
+            AccountViewModel(appPreferencesService, authenticationUseCase, accountDetailsUseCase)
         testDispatcher.scheduler.advanceUntilIdle()
 
         // Then
@@ -288,7 +293,8 @@ class   AccountViewModelTest {
         coEvery { accountDetailsUseCase.invoke() } returns AccountInfo(1, "", "")
 
         // When
-        viewModel = AccountViewModel(appPreferencesService, authenticationUseCase, accountDetailsUseCase)
+        viewModel =
+            AccountViewModel(appPreferencesService, authenticationUseCase, accountDetailsUseCase)
         testDispatcher.scheduler.advanceUntilIdle()
         
         contentRestrictionFlow.value = ContentRestrictionLevel.STRICT
@@ -301,8 +307,9 @@ class   AccountViewModelTest {
     private fun setupViewModel() {
         coEvery { authenticationUseCase.isLoggedIn() } returns true
         coEvery { accountDetailsUseCase.invoke() } returns AccountInfo(1, "TestUser", "avatar.jpg")
-        
-        viewModel = AccountViewModel(appPreferencesService, authenticationUseCase, accountDetailsUseCase)
+
+        viewModel =
+            AccountViewModel(appPreferencesService, authenticationUseCase, accountDetailsUseCase)
         testDispatcher.scheduler.advanceUntilIdle()
     }
 }
